Functioning spatially varying and invariant convolutions
Boxfilter, spatiallyInvariantGaussianKernel, and analyticKernel
roughly function on masked fits images.  The 2 edge pixels of the variance
and mask planes do not match the LSST output.  It makes sense that
the mask output should be computed as the bitwise OR of each input
mask whose kernel value is non-zero, but the LSST reference code
appears to take the bitwise OR of all input mask values, regardless
of whether their corresponding kernel values are zero.