Linear combination kernel working
This version includes failed attempts to improve the efficiency
of linear combination by reordering the computation.  However,
the variance cannot be computed efficiently using the loop reordering.
Computation of the mask output doesn't work either, if input mask
pixels must be ORed on the condition that the kernel is zero
(or close to zero).